Phase 3: Remove all old facade and builder code
This commit completes the migration by removing all old code:

1. **Deleted old files** (868 lines removed):
   - ast_construction.rs (old facade layer)
   - builders.rs (old text-based builders)
   - ast_construction_tests.rs (tests for old facade)

2. **Added text-based API to ast_builder.rs** (230 lines added):
   - build_paragraph_from_text()
   - build_session_from_text()
   - build_definition_from_text()
   - build_annotation_from_text()
   - build_list_from_items()
   - build_foreign_block_from_text()
   These are needed by the reference parser which works with
   pre-extracted text during combinator parsing.

3. **Updated all parsers to use new API**:
   - Reference parser: use ast_builder::*_from_text functions
   - Linebased parser: use ast_builder functions directly
   - Engine: use ast_builder::build_session_from_text

4. **Updated mod.rs**:
   - Removed old exports (builders, ast_construction)
   - Clean module structure with only new API

**Architecture is now unified**:
- Single source of truth: ast_builder.rs
- Three-layer pipeline: normalize → extract → create
- Two API tiers:
  * Token-based (for parsers with LineToken/tokens)
  * Text-based (for parsers that extract text early)

**Results**:
- Net deletion of 605 lines
- All 386 tests passing
- Foreign block indentation wall bug fixed
- No old/deprecated code remaining
