+++ src/lex/parsers/common/mod.rs
@@ -3,22 +3,13 @@
 //! This module contains shared interfaces and utilities for parser implementations.
 
 pub mod ast_builder;
-pub mod ast_construction;
 pub mod ast_creation;
-pub mod builders;
 pub mod data_extraction;
 pub mod interface;
 pub mod location;
 pub mod token_normalization;
 pub mod token_processing;
 
-#[cfg(test)]
-mod ast_construction_tests;
-
-pub use builders::{
-    build_annotation, build_definition, build_foreign_block, build_list, build_list_item,
-    build_paragraph, build_session, extract_text_from_span,
-};
 pub use interface::{
     LineBasedParserImpl, ParseError, Parser, ParserInput, ParserRegistry, ReferenceParserImpl,
 };
